+++ .github/workflows/e2e.yaml
@@ -17,16 +17,23 @@ jobs:
         with:
           python-version: "3.12"
       - name: setup node
-        uses: actions/setup-node@v3
+        uses: actions/setup-node@v4
         with:
-          node-version: "16"
+          node-version: "22"
 
-      - name: cache
+      - name: cache1
         uses: actions/cache@v3
         with:
           path: |
             node_modules
-          key: ${{ runner.os }}-${{ hashFiles('**/yarn.lock') }}
+          key: ${{ runner.os }}-${{ hashFiles('./yarn.lock') }}
+
+      - name: cache2
+        uses: actions/cache@v3
+        with:
+          path: |
+            storybook/node_modules
+          key: ${{ runner.os }}-${{ hashFiles('./storybook/yarn.lock') }}
 
       - name: install
         run: |
@@ -34,11 +41,14 @@ jobs:
           npx playwright install --with-deps
 
       - name: build
-        run: yarn run build-storybook
+        run: |
+          cd storybook
+          yarn install
+          yarn run build
 
       - name: run server
         run: |
-          python -m http.server 5233 --directory "${{ github.workspace }}/storybook-static" &
+          python -m http.server 5233 --directory "${{ github.workspace }}/storybook/storybook-static" &
 
       - name: e2e test
         run: yarn e2e
