+++ package.json
@@ -7,8 +7,8 @@
   "scripts": {
     "build": "rm -rf ./dist || true && tsc --project ./src/index.ts -p ./tsconfig.json",
     "audit-fix": "yarn-audit-fix --force",
-    "storybook": "start-storybook -p 5233",
-    "build-storybook": "build-storybook",
+    "storybook:run": "cd storybook && yarn run run",
+    "storybook:build": "cd storybook && yarn run build",
     "less": "lessc --clean-css ./src/styles/root.less ./src/styles/root.min.css",
     "generate-style": "node ./generate-style.js",
     "playwright": "npx playwright test --ui",
@@ -45,7 +45,7 @@
     "@babel/preset-env": "^7.14.4",
     "@babel/preset-react": "^7.13.13",
     "@babel/preset-typescript": "^7.13.0",
-    "@playwright/test": "^1.41.1",
+    "@playwright/test": "^1.50.1",
     "@storybook/addon-essentials": "^6.5.12",
     "@storybook/addon-links": "^6.5.12",
     "@storybook/react": "^6.5.12",
@@ -63,7 +63,6 @@
     "jest-environment-jsdom": "^29.4.3",
     "less": "^4.1.3",
     "less-plugin-clean-css": "^1.5.1",
-    "playwright": "^1.50.1",
     "prettier": "^3.2.5",
     "react": ">=16.9.0",
     "react-dom": ">=16.9.0",
