+++ storybook/examples/events/onchange.stories.tsx
@@ -1,8 +1,5 @@
 import React from "react";
-import { GridSheet } from "../../../src";
-import { constructInitialCells } from "../../../src/lib/structs";
-import { createTableRef } from "../../../src/components/Tabular";
-import {HistoryType} from "../../../src/types";
+import { GridSheet, constructInitialCells, createTableRef, HistoryType } from "@gridsheet/react-core";
 
 export default {
   title: "Table operations",
