+++ storybook/examples/protection/protection.stories.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { ComponentStory } from "@storybook/react";
-import { constructInitialCells, GridSheet, prevention } from "../../../src";
+import { constructInitialCells, GridSheet, prevention } from "@gridsheet/react-core";
 
 export default {
   title: "Basic",
