+++ storybook/package.json
@@ -1,0 +1,62 @@
+{
+  "name": "gridsheet-storybook",
+  "version": "1.0.0",
+  "main": "dist/index.js",
+  "module": "dist/index.js",
+  "scripts": {
+    "run": "start-storybook -p 5233",
+    "build": "build-storybook"
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.14.3",
+    "@babel/core": "^7.14.3",
+    "@babel/preset-env": "^7.14.4",
+    "@babel/preset-react": "^7.13.13",
+    "@babel/preset-typescript": "^7.13.0",
+    "@gridsheet/react-core": "file:../src",
+    "@storybook/addon-essentials": "^6.5.12",
+    "@storybook/addon-links": "^6.5.12",
+    "@storybook/react": "^6.5.12",
+    "@types/date-fns": "^2.6.0",
+    "@types/jest": "^29.4.0",
+    "@types/react": "^16.9.24",
+    "@types/react-dom": "^16.9.24",
+    "@types/storybook__react": "^5.2.1",
+    "@typescript-eslint/parser": "^4.14.2",
+    "babel-loader": "^8.2.2",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "jest": "^29.4.3",
+    "jest-environment-jsdom": "^29.4.3",
+    "less": "^4.1.3",
+    "less-plugin-clean-css": "^1.5.1",
+    "prettier": "^3.2.5",
+    "react": ">=16.9.0",
+    "react-dom": ">=16.9.0",
+    "react-is": "^16.13.1",
+    "require-context.macro": "^1.2.2",
+    "storybook": "^6.5.12",
+    "ts-jest": "^29.0.5",
+    "ts-loader": "^8.4.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.0.3",
+    "typescript-eslint": "^7.0.2",
+    "webpack-cli": "^4.7.0",
+    "yarn-audit-fix": "^9.3.6"
+  },
+  "resolutions": {
+    "trim": "^0.0.3",
+    "trim-newlines": "^3.0.1",
+    "got": "^11.8.5",
+    "glob-parent": "^5.1.2",
+    "babel-plugin-macros": "^3.1.0"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "dependencies": {
+    "@storybook/react-webpack5": "^8.4.5"
+  }
+}
